refactor: Phase C1 - extract Collection headers and Settings sections
CollectionPage (615→246 lines):
- CollectionPwaHeader: PWA compact search/filter/sort hamburger menu
- CollectionDesktopHeader: Desktop sticky toolbar with filters

SettingsPage (1341→433 lines):
- SettingsAccountSection: avatar, profile, NumisBids, password, WebAuthn
- SettingsDataSection: API keys, tags, import/export

All components self-contained with typed props/emits. vue-tsc passes clean.
