+++ src/web/src/pages/SettingsPage.vue
@@ -39,164 +39,7 @@
       </div>
 
       <!-- Account Tab -->
-      <section v-if="activeTab === 'account'" class="settings-section card">
-        <h2>Account</h2>
-
-        <!-- Avatar -->
-        <div class="setting-item avatar-section">
-          <div class="avatar-preview">
-            <img :src="avatarUrl" alt="Avatar" class="avatar-img" />
-          </div>
-          <div class="avatar-actions">
-            <label class="btn btn-secondary btn-sm">
-              Upload Avatar
-              <input type="file" accept="image/*" hidden @change="handleAvatarUpload" />
-            </label>
-            <button v-if="auth.user?.avatarPath" class="btn btn-danger btn-sm" @click="handleAvatarDelete">Remove</button>
-          </div>
-        </div>
-
-        <div class="setting-item">
-          <div class="setting-info">
-            <span class="setting-label">Username</span>
-            <span class="setting-value">{{ auth.user?.username }}</span>
-          </div>
-        </div>
-        <div class="setting-item">
-          <div class="setting-info">
-            <span class="setting-label">Role</span>
-            <span class="setting-value badge" :class="`badge-${auth.user?.role === 'admin' ? 'roman' : 'modern'}`">
-              {{ auth.user?.role }}
-            </span>
-          </div>
-        </div>
-
-        <!-- Profile / Social Settings -->
-        <h3>Profile</h3>
-        <div class="form-group">
-          <label class="form-label">Email</label>
-          <input v-model="profileEmail" type="email" class="form-input" placeholder="you@example.com" />
-        </div>
-        <div class="form-group">
-          <label class="form-label">Bio</label>
-          <input v-model="profileBio" class="form-input" placeholder="Tell collectors about yourself..." maxlength="200" />
-        </div>
-        <div class="form-group">
-          <label class="form-label">ZIP Code</label>
-          <input v-model="profileZipCode" class="form-input" placeholder="e.g. 90210" maxlength="10" />
-          <span class="setting-desc" style="font-size: 0.8rem; margin-top: 0.25rem; display: block">Used by the Agent to find nearby coin shows and dealers</span>
-        </div>
-
-        <h3>NumisBids Integration</h3>
-        <p class="setting-desc" style="margin-bottom: 0.75rem">
-          Connect your NumisBids account to sync your watchlist and track auction lots.
-        </p>
-        <div class="form-group">
-          <label class="form-label">NumisBids Username</label>
-          <input v-model="nbUsername" class="form-input" placeholder="Your NumisBids username" autocomplete="off" />
-        </div>
-        <div class="form-group">
-          <label class="form-label">NumisBids Password</label>
-          <input v-model="nbPassword" type="password" class="form-input" placeholder="Your NumisBids password" autocomplete="new-password" />
-          <span class="setting-desc" style="font-size: 0.8rem; margin-top: 0.25rem; display: block">Stored securely on the server. Used only for watchlist sync.</span>
-        </div>
-        <div v-if="nbValidating" class="nb-status validating">
-          Validating NumisBids credentials...
-        </div>
-        <div v-else-if="nbValidationError" class="nb-status error">
-          {{ nbValidationError }}
-        </div>
-        <div v-else-if="auth.user?.numisBidsConfigured" class="nb-status connected">
-          NumisBids account connected
-        </div>
-        <div class="setting-item">
-          <div class="setting-info">
-            <span class="setting-label">Public Collection</span>
-            <span class="setting-desc">Allow other users to follow you and view your coins</span>
-          </div>
-          <label class="toggle">
-            <input type="checkbox" :checked="profilePublic" @change="onPublicToggle" />
-            <span class="toggle-slider"></span>
-          </label>
-        </div>
-        <button class="btn btn-primary btn-sm" @click="handleSaveProfile" :disabled="profileSaving || nbValidating" style="margin-top: 0.5rem">
-          {{ nbValidating ? 'Validating...' : profileSaving ? 'Saving...' : 'Save Profile' }}
-        </button>
-        <p v-if="profileMsg" class="msg" :class="{ error: profileError }" style="margin-top: 0.5rem">{{ profileMsg }}</p>
-
-    <!-- Privacy Warning Modal -->
-    <Teleport to="body">
-      <div v-if="showPrivacyWarning" class="modal-overlay" @click.self="cancelGoPrivate">
-        <div class="modal-content card" style="max-width: 440px;">
-          <div class="modal-header">
-            <h2 style="display: flex; align-items: center; gap: 0.5rem; margin: 0; font-size: 1rem;">
-              ⚠️ Make Collection Private?
-            </h2>
-          </div>
-          <div class="modal-body" style="padding: 1.25rem;">
-            <p style="color: var(--text-secondary); line-height: 1.5; margin: 0 0 0.75rem;">
-              Setting your profile to private will <strong style="color: var(--text-primary);">permanently remove all your followers</strong>.
-              They will need to send new follow requests if you make your profile public again.
-            </p>
-            <p style="color: var(--text-secondary); line-height: 1.5; margin: 0 0 1rem;">
-              You will also be hidden from user search results.
-            </p>
-            <div style="display: flex; gap: 0.75rem; justify-content: flex-end;">
-              <button class="btn btn-secondary btn-sm" @click="cancelGoPrivate">Cancel</button>
-              <button class="btn btn-danger btn-sm" @click="confirmGoPrivate">Make Private</button>
-            </div>
-          </div>
-        </div>
-      </div>
-    </Teleport>
-
-        <h3>Change Password</h3>
-        <form class="password-form" @submit.prevent="handleChangePassword">
-          <div class="form-group">
-            <label class="form-label">Current Password</label>
-            <input v-model="currentPassword" type="password" class="form-input" required />
-          </div>
-          <div class="form-group">
-            <label class="form-label">New Password</label>
-            <input v-model="newPassword" type="password" class="form-input" required minlength="6" />
-          </div>
-          <div class="form-group">
-            <label class="form-label">Confirm New Password</label>
-            <input v-model="confirmPassword" type="password" class="form-input" required />
-          </div>
-          <p v-if="passwordMsg" class="msg" :class="{ error: passwordError }">{{ passwordMsg }}</p>
-          <button type="submit" class="btn btn-primary btn-sm" :disabled="passwordLoading">
-            {{ passwordLoading ? 'Changing...' : 'Change Password' }}
-          </button>
-        </form>
-
-        <template v-if="supportsWebAuthn">
-          <h3>Biometric Login</h3>
-          <p class="setting-desc" style="margin-bottom: 0.75rem">
-            Register Face ID, Touch ID, or fingerprint for quick sign-in on this device.
-          </p>
-
-          <button
-            class="btn btn-primary btn-sm"
-            :disabled="registeringCredential"
-            @click="handleRegisterCredential"
-          >
-            {{ registeringCredential ? 'Registering...' : '🔐 Register Biometric' }}
-          </button>
-          <p v-if="credentialMsg" class="msg" :class="{ error: credentialError }" style="margin-top: 0.5rem">{{ credentialMsg }}</p>
-
-          <div v-if="webauthnCredentials.length" class="apikey-list">
-            <div v-for="cred in webauthnCredentials" :key="cred.id" class="apikey-item">
-              <div class="apikey-item-info">
-                <span class="apikey-item-name">{{ cred.name }}</span>
-                <span class="apikey-item-meta">Registered {{ formatDate(cred.createdAt) }}</span>
-              </div>
-              <button class="btn btn-danger btn-sm" @click="handleDeleteCredential(cred.id)">Remove</button>
-            </div>
-          </div>
-          <p v-else-if="!registeringCredential" class="setting-desc" style="margin-top: 0.5rem">No biometric credentials registered.</p>
-        </template>
-      </section>
+      <SettingsAccountSection v-if="activeTab === 'account'" ref="accountSection" />
 
       <!-- Appearance Tab -->
       <SettingsAppearanceSection
@@ -213,155 +56,7 @@
       />
 
       <!-- Data Tab -->
-      <section v-if="activeTab === 'data'" class="settings-section card">
-        <h2>Data Management</h2>
-        <div class="setting-item">
-          <div class="setting-info">
-            <span class="setting-label">Export Collection</span>
-            <span class="setting-desc">Download your collection data and photos as a zip archive</span>
-          </div>
-          <button class="btn btn-secondary btn-sm" :disabled="exporting" @click="handleExport">
-            {{ exporting ? 'Exporting...' : 'Export ZIP' }}
-          </button>
-        </div>
-        <div class="setting-item">
-          <div class="setting-info">
-            <span class="setting-label">PDF Catalog</span>
-            <span class="setting-desc">Generate a styled PDF catalog with photos, grades, and valuations</span>
-          </div>
-          <button class="btn btn-secondary btn-sm" :disabled="exportingPdf" @click="handleExportPDF">
-            {{ exportingPdf ? 'Generating...' : 'Export PDF' }}
-          </button>
-        </div>
-        <div class="setting-item">
-          <div class="setting-info">
-            <span class="setting-label">Import Collection</span>
-            <span class="setting-desc">Import coins from a JSON file</span>
-          </div>
-          <label class="btn btn-secondary btn-sm import-btn">
-            📤 Import
-            <input type="file" accept=".json" hidden @change="handleImport" />
-          </label>
-        </div>
-        <p v-if="dataMsg" class="msg" :class="{ error: dataError }">{{ dataMsg }}</p>
-
-        <h3>API Keys</h3>
-        <p class="setting-desc" style="margin-bottom: 1rem">
-          Generate API keys to access your collection from external tools and scripts. Use the <code>X-API-Key</code> header to authenticate.
-        </p>
-
-        <div class="apikey-generate">
-          <input
-            v-model="apiKeyName"
-            type="text"
-            class="form-input"
-            placeholder="Key name (e.g. My Script)"
-            :disabled="generatingKey"
-          />
-          <button
-            class="btn btn-primary btn-sm"
-            :disabled="!apiKeyName.trim() || generatingKey"
-            @click="handleGenerateKey"
-          >
-            {{ generatingKey ? 'Generating...' : '🔑 Generate Key' }}
-          </button>
-        </div>
-
-        <div v-if="newlyGeneratedKey" class="apikey-reveal">
-          <p class="apikey-reveal-warning">
-            ⚠️ Copy this key now — it will not be shown again.
-          </p>
-          <div class="apikey-reveal-box">
-            <code class="apikey-reveal-value">{{ newlyGeneratedKey }}</code>
-            <button class="btn btn-secondary btn-sm" @click="copyKey">
-              {{ keyCopied ? '✓ Copied' : '📋 Copy' }}
-            </button>
-          </div>
-        </div>
-
-        <p v-if="apiKeyMsg" class="msg" :class="{ error: apiKeyError }">{{ apiKeyMsg }}</p>
-
-        <div v-if="apiKeys.length" class="apikey-list">
-          <div
-            v-for="key in apiKeys"
-            :key="key.id"
-            class="apikey-item"
-            :class="{ revoked: key.revokedAt }"
-          >
-            <div class="apikey-item-info">
-              <span class="apikey-item-name">{{ key.name }}</span>
-              <span class="apikey-item-meta">
-                ...{{ key.keyPrefix }}
-                · Created {{ formatDate(key.createdAt) }}
-                <template v-if="key.lastUsedAt"> · Last used {{ formatDate(key.lastUsedAt) }}</template>
-              </span>
-            </div>
-            <span v-if="key.revokedAt" class="apikey-item-badge revoked-badge">Revoked</span>
-            <button
-              v-else
-              class="btn btn-danger btn-sm"
-              @click="handleRevokeKey(key.id)"
-            >
-              Revoke
-            </button>
-          </div>
-        </div>
-        <p v-else-if="!generatingKey" class="setting-desc" style="margin-top: 0.5rem">No API keys yet.</p>
-
-        <h3 style="margin-top: 2rem">Tags</h3>
-        <p class="setting-desc">Create custom tags to organize and filter your coins.</p>
-
-        <div class="tag-create-form">
-          <input
-            v-model="newTagName"
-            type="text"
-            class="form-input"
-            placeholder="New tag name..."
-            maxlength="50"
-            @keydown.enter="handleCreateTag"
-          />
-          <div class="tag-color-picker">
-            <button
-              v-for="c in TAG_COLORS"
-              :key="c"
-              class="color-swatch"
-              :class="{ active: newTagColor === c }"
-              :style="{ backgroundColor: c }"
-              @click="newTagColor = c"
-            ></button>
-          </div>
-          <button class="btn btn-primary btn-sm" @click="handleCreateTag" :disabled="!newTagName.trim()">Create Tag</button>
-        </div>
-        <p v-if="tagError" class="tag-error">{{ tagError }}</p>
-
-        <div v-if="tagList.length" class="tag-list">
-          <div v-for="tag in tagList" :key="tag.id" class="tag-list-item">
-            <template v-if="editingTag?.id === tag.id">
-              <input v-model="editTagName" class="form-input tag-edit-input" maxlength="50" @keydown.enter="handleSaveTag" />
-              <div class="tag-color-picker">
-                <button
-                  v-for="c in TAG_COLORS"
-                  :key="c"
-                  class="color-swatch sm"
-                  :class="{ active: editTagColor === c }"
-                  :style="{ backgroundColor: c }"
-                  @click="editTagColor = c"
-                ></button>
-              </div>
-              <button class="btn btn-primary btn-sm" @click="handleSaveTag">Save</button>
-              <button class="btn btn-secondary btn-sm" @click="editingTag = null">Cancel</button>
-            </template>
-            <template v-else>
-              <span class="tag-preview" :style="{ backgroundColor: tag.color + '22', color: tag.color, borderColor: tag.color + '44' }">{{ tag.name }}</span>
-              <div class="tag-actions">
-                <button class="btn btn-secondary btn-sm" @click="startEditTag(tag)">Edit</button>
-                <button class="btn btn-danger btn-sm" @click="handleDeleteTag(tag)">Delete</button>
-              </div>
-            </template>
-          </div>
-        </div>
-        <p v-else class="empty-tags">No tags created yet. Create your first tag above.</p>
-      </section>
+      <SettingsDataSection v-if="activeTab === 'data'" ref="dataSection" />
 
       <!-- Tools Tab -->
       <SettingsToolsSection
@@ -401,26 +96,21 @@ import { useRoute, useRouter } from 'vue-router'
 import { useAuthStore } from '@/stores/auth'
 import PullToRefresh from '@/components/PullToRefresh.vue'
 import {
-  exportCollection, exportCatalogPDF, importCollection,
-  generateApiKey, listApiKeys, revokeApiKey,
-  webauthnRegisterBegin, webauthnRegisterFinish,
-  webauthnListCredentials, webauthnDeleteCredential,
   listConversations, getConversation, deleteConversation,
   getBlockedUsers, unblockFollower,
-  getTags, createTag, updateTag as updateTagApi, deleteTag,
 } from '@/api/client'
 import type { ConversationSummary } from '@/api/client'
 import { useDialog } from '@/composables/useDialog'
 import { usePwa } from '@/composables/usePwa'
-import { useSettingsProfile } from '@/composables/useSettingsProfile'
-import type { Coin, Theme, ApiKey, WebAuthnCredentialInfo, Tag } from '@/types'
+import type { Theme } from '@/types'
 import CoinSearchChat from '@/components/CoinSearchChat.vue'
-import ImageProcessor from '@/components/ImageProcessor.vue'
 import HelpSection from '@/components/HelpSection.vue'
+import SettingsAccountSection from '@/components/settings/SettingsAccountSection.vue'
 import SettingsAppearanceSection from '@/components/settings/SettingsAppearanceSection.vue'
+import SettingsDataSection from '@/components/settings/SettingsDataSection.vue'
 import SavedConversationsSection from '@/components/settings/SavedConversationsSection.vue'
 import SettingsToolsSection from '@/components/settings/SettingsToolsSection.vue'
-import { User, Palette, Database, MessageSquare, HelpCircle, Wrench, Menu, ShieldCheck, Tags } from 'lucide-vue-next'
+import { User, Palette, Database, MessageSquare, HelpCircle, Wrench, Menu, ShieldCheck } from 'lucide-vue-next'
 
 const tabIcons: Record<string, Component> = {
   account: User,
@@ -441,15 +131,8 @@ const route = useRoute()
 const router = useRouter()
 const auth = useAuthStore()
 
-const {
-  avatarUrl, handleAvatarUpload, handleAvatarDelete,
-  profileEmail, profileBio, profileZipCode,
-  nbUsername, nbPassword, profilePublic, profileMsg, profileError, profileSaving,
-  showPrivacyWarning, onPublicToggle, confirmGoPrivate, cancelGoPrivate,
-  nbValidating, nbValidationError, handleSaveProfile,
-  currentPassword, newPassword, confirmPassword,
-  passwordMsg, passwordError, passwordLoading, handleChangePassword,
-} = useSettingsProfile()
+const accountSection = ref<InstanceType<typeof SettingsAccountSection> | null>(null)
+const dataSection = ref<InstanceType<typeof SettingsDataSection> | null>(null)
 
 const baseTabs = [
   { id: 'account', label: 'Account' },
@@ -516,75 +199,6 @@ async function handleUnblock(user: { id: number; username: string; avatarPath: s
   }
 }
 
-// Tag management
-const tagList = ref<Tag[]>([])
-const newTagName = ref('')
-const newTagColor = ref('#6b7280')
-const editingTag = ref<Tag | null>(null)
-const editTagName = ref('')
-const editTagColor = ref('')
-const tagError = ref('')
-
-const TAG_COLORS = ['#6b7280', '#ef4444', '#f59e0b', '#10b981', '#3b82f6', '#8b5cf6', '#ec4899', '#14b8a6', '#f97316', '#6366f1']
-
-async function loadTags() {
-  try {
-    const res = await getTags()
-    tagList.value = res.data?.tags ?? []
-  } catch { tagList.value = [] }
-}
-
-async function handleCreateTag() {
-  tagError.value = ''
-  const name = newTagName.value.trim()
-  if (!name) return
-  try {
-    await createTag({ name, color: newTagColor.value })
-    newTagName.value = ''
-    newTagColor.value = '#6b7280'
-    await loadTags()
-  } catch (e: unknown) {
-    if (typeof e === 'object' && e !== null && 'response' in e) {
-      const axiosErr = e as { response?: { data?: { error?: string } } }
-      tagError.value = axiosErr.response?.data?.error ?? 'Failed to create tag'
-    } else {
-      tagError.value = 'Failed to create tag'
-    }
-  }
-}
-
-function startEditTag(tag: Tag) {
-  editingTag.value = tag
-  editTagName.value = tag.name
-  editTagColor.value = tag.color
-}
-
-async function handleSaveTag() {
-  tagError.value = ''
-  if (!editingTag.value) return
-  try {
-    await updateTagApi(editingTag.value.id, { name: editTagName.value.trim(), color: editTagColor.value })
-    editingTag.value = null
-    await loadTags()
-  } catch (e: unknown) {
-    if (typeof e === 'object' && e !== null && 'response' in e) {
-      const axiosErr = e as { response?: { data?: { error?: string } } }
-      tagError.value = axiosErr.response?.data?.error ?? 'Failed to update tag'
-    } else {
-      tagError.value = 'Failed to update tag'
-    }
-  }
-}
-
-async function handleDeleteTag(tag: Tag) {
-  const confirmed = await showConfirm(`Delete tag "${tag.name}"? It will be removed from all coins.`, { title: 'Delete Tag', variant: 'danger' })
-  if (!confirmed) return
-  try {
-    await deleteTag(tag.id)
-    await loadTags()
-  } catch { /* ignore */ }
-}
-
 // Theme
 const theme = ref<Theme>((localStorage.getItem('theme') as Theme) || 'dark')
 
@@ -617,233 +231,6 @@ const defaultSort = ref(localStorage.getItem('defaultSort') || 'updated_at_desc'
 
 function saveDefaultSort() {
   localStorage.setItem('defaultSort', defaultSort.value)
-}
-
-// Data
-const exporting = ref(false)
-const exportingPdf = ref(false)
-const dataMsg = ref('')
-const dataError = ref(false)
-
-async function handleExport() {
-  exporting.value = true
-  dataMsg.value = ''
-  try {
-    const res = await exportCollection()
-    const blob = new Blob([res.data], { type: 'application/zip' })
-    const url = URL.createObjectURL(blob)
-    const a = document.createElement('a')
-    a.href = url
-    a.download = `ancient-coins-export-${new Date().toISOString().slice(0, 10)}.zip`
-    a.click()
-    URL.revokeObjectURL(url)
-    dataMsg.value = 'Export downloaded'
-  } catch {
-    dataMsg.value = 'Export failed'
-    dataError.value = true
-  } finally {
-    exporting.value = false
-  }
-}
-
-async function handleExportPDF() {
-  exportingPdf.value = true
-  dataMsg.value = ''
-  try {
-    const res = await exportCatalogPDF()
-    const blob = new Blob([res.data], { type: 'application/pdf' })
-    const url = URL.createObjectURL(blob)
-    const a = document.createElement('a')
-    a.href = url
-    a.download = `coin-catalog-${new Date().toISOString().slice(0, 10)}.pdf`
-    a.click()
-    URL.revokeObjectURL(url)
-    dataMsg.value = 'PDF catalog downloaded'
-  } catch {
-    dataMsg.value = 'PDF generation failed'
-    dataError.value = true
-  } finally {
-    exportingPdf.value = false
-  }
-}
-
-async function handleImport(e: Event) {
-  const file = (e.target as HTMLInputElement).files?.[0]
-  if (!file) return
-
-  dataMsg.value = ''
-  dataError.value = false
-
-  try {
-    const text = await file.text()
-    const coins: Coin[] = JSON.parse(text)
-    const res = await importCollection(coins)
-    dataMsg.value = `Imported ${res.data.imported} coins`
-  } catch {
-    dataMsg.value = 'Import failed — ensure valid JSON format'
-    dataError.value = true
-  }
-}
-
-// API Keys
-const apiKeys = ref<ApiKey[]>([])
-const apiKeyName = ref('')
-const newlyGeneratedKey = ref('')
-const keyCopied = ref(false)
-const generatingKey = ref(false)
-const apiKeyMsg = ref('')
-const apiKeyError = ref(false)
-
-async function loadApiKeys() {
-  try {
-    const res = await listApiKeys()
-    apiKeys.value = res.data
-  } catch {
-    // silently fail on load
-  }
-}
-
-async function handleGenerateKey() {
-  if (!apiKeyName.value.trim()) return
-
-  generatingKey.value = true
-  apiKeyMsg.value = ''
-  apiKeyError.value = false
-  newlyGeneratedKey.value = ''
-  keyCopied.value = false
-
-  try {
-    const res = await generateApiKey(apiKeyName.value.trim())
-    newlyGeneratedKey.value = res.data.key
-    apiKeyName.value = ''
-    await loadApiKeys()
-  } catch {
-    apiKeyMsg.value = 'Failed to generate API key'
-    apiKeyError.value = true
-  } finally {
-    generatingKey.value = false
-  }
-}
-
-async function copyKey() {
-  try {
-    await navigator.clipboard.writeText(newlyGeneratedKey.value)
-    keyCopied.value = true
-    setTimeout(() => { keyCopied.value = false }, 3000)
-  } catch {
-    // Fallback for non-HTTPS contexts
-    const textarea = document.createElement('textarea')
-    textarea.value = newlyGeneratedKey.value
-    document.body.appendChild(textarea)
-    textarea.select()
-    document.execCommand('copy')
-    document.body.removeChild(textarea)
-    keyCopied.value = true
-    setTimeout(() => { keyCopied.value = false }, 3000)
-  }
-}
-
-async function handleRevokeKey(id: number) {
-  apiKeyMsg.value = ''
-  apiKeyError.value = false
-  try {
-    await revokeApiKey(id)
-    await loadApiKeys()
-    newlyGeneratedKey.value = ''
-  } catch {
-    apiKeyMsg.value = 'Failed to revoke key'
-    apiKeyError.value = true
-  }
-}
-
-function formatDate(dateStr: string) {
-  return new Date(dateStr).toLocaleDateString(undefined, {
-    year: 'numeric', month: 'short', day: 'numeric',
-  })
-}
-
-// WebAuthn Biometric
-const supportsWebAuthn = !!window.PublicKeyCredential
-const webauthnCredentials = ref<WebAuthnCredentialInfo[]>([])
-const registeringCredential = ref(false)
-const credentialMsg = ref('')
-const credentialError = ref(false)
-
-async function loadCredentials() {
-  try {
-    const res = await webauthnListCredentials()
-    webauthnCredentials.value = res.data
-  } catch {
-    // silently fail
-  }
-}
-
-function base64urlToBuffer(base64url: string): ArrayBuffer {
-  const base64 = base64url.replace(/-/g, '+').replace(/_/g, '/')
-  const pad = base64.length % 4 === 0 ? '' : '='.repeat(4 - (base64.length % 4))
-  const binary = atob(base64 + pad)
-  const bytes = new Uint8Array(binary.length)
-  for (let i = 0; i < binary.length; i++) bytes[i] = binary.charCodeAt(i)
-  return bytes.buffer
-}
-
-async function handleRegisterCredential() {
-  registeringCredential.value = true
-  credentialMsg.value = ''
-  credentialError.value = false
-
-  try {
-    // Begin registration — get options from server
-    const beginRes = await webauthnRegisterBegin()
-    const options = beginRes.data
-
-    // Convert base64url fields to ArrayBuffers for the browser API
-    const publicKeyOptions: PublicKeyCredentialCreationOptions = {
-      challenge: base64urlToBuffer(options.publicKey.challenge),
-      rp: options.publicKey.rp,
-      user: {
-        id: base64urlToBuffer(options.publicKey.user.id),
-        name: options.publicKey.user.name,
-        displayName: options.publicKey.user.displayName,
-      },
-      pubKeyCredParams: options.publicKey.pubKeyCredParams,
-      timeout: options.publicKey.timeout || 60000,
-      authenticatorSelection: options.publicKey.authenticatorSelection,
-      attestation: options.publicKey.attestation || 'none',
-      excludeCredentials: (options.publicKey.excludeCredentials || []).map((c: { id: string; type: string; transports?: string[] }) => ({
-        id: base64urlToBuffer(c.id),
-        type: c.type,
-        transports: c.transports,
-      })),
-    }
-
-    // Call browser WebAuthn API (triggers Face ID / fingerprint prompt)
-    const credential = await navigator.credentials.create({
-      publicKey: publicKeyOptions,
-    }) as PublicKeyCredential
-
-    // Finish registration — send attestation to server
-    await webauthnRegisterFinish(credential)
-
-    credentialMsg.value = 'Biometric credential registered!'
-    await loadCredentials()
-  } catch (e: unknown) {
-    credentialMsg.value = e instanceof Error ? e.message : 'Registration failed'
-    credentialError.value = true
-  } finally {
-    registeringCredential.value = false
-  }
-}
-
-async function handleDeleteCredential(id: number) {
-  if (!await showConfirm('Remove this biometric credential?', { title: 'Remove Credential' })) return
-  try {
-    await webauthnDeleteCredential(id)
-    await loadCredentials()
-  } catch {
-    credentialMsg.value = 'Failed to remove credential'
-    credentialError.value = true
-  }
 }
 
 // Saved Conversations
@@ -890,47 +277,20 @@ async function handleDeleteConversation(id: number) {
 
 async function handleRefresh() {
   await Promise.all([
-    loadApiKeys(),
+    dataSection.value?.loadApiKeys() ?? Promise.resolve(),
     loadConversations(),
     loadBlockedUsers(),
-    supportsWebAuthn ? loadCredentials() : Promise.resolve(),
+    accountSection.value?.supportsWebAuthn ? accountSection.value.loadCredentials() : Promise.resolve(),
   ])
 }
 
 onMounted(() => {
-  loadApiKeys()
   loadConversations()
   loadBlockedUsers()
-  loadTags()
-  if (supportsWebAuthn) loadCredentials()
 })
 </script>
 
 <style scoped>
-.avatar-section {
-  display: flex;
-  align-items: center;
-  gap: 1rem;
-}
-
-.avatar-preview {
-  flex-shrink: 0;
-}
-
-.avatar-img {
-  width: 64px;
-  height: 64px;
-  border-radius: 50%;
-  object-fit: cover;
-  border: 2px solid var(--accent-gold-dim, #c9a84c);
-}
-
-.avatar-actions {
-  display: flex;
-  gap: 0.5rem;
-  flex-wrap: wrap;
-}
-
 .settings-layout {
   max-width: 800px;
   margin-left: auto;
@@ -1018,42 +378,9 @@ onMounted(() => {
   color: var(--text-muted);
 }
 
-.nb-status {
-  font-size: 0.82rem;
-  padding: 0.4rem 0.75rem;
-  border-radius: var(--radius-sm);
-  margin-top: 0.25rem;
-}
-
-.nb-status.connected {
-  background: rgba(74, 222, 128, 0.1);
-  color: #4ade80;
-  border: 1px solid rgba(74, 222, 128, 0.2);
-}
-
-.nb-status.validating {
-  background: rgba(250, 204, 21, 0.1);
-  color: #facc15;
-  border: 1px solid rgba(250, 204, 21, 0.2);
-}
-
-.nb-status.error {
-  background: rgba(239, 68, 68, 0.1);
-  color: #ef4444;
-  border: 1px solid rgba(239, 68, 68, 0.2);
-}
-
 .setting-value {
   font-size: 0.9rem;
   color: var(--text-secondary);
-}
-
-.password-form {
-  max-width: 350px;
-}
-
-.import-btn {
-  cursor: pointer;
 }
 
 .msg {
@@ -1064,98 +391,6 @@ onMounted(() => {
 
 .msg.error {
   color: #e74c3c;
-}
-
-.apikey-generate {
-  display: flex;
-  gap: 0.75rem;
-  align-items: center;
-  margin-bottom: 0.75rem;
-}
-
-.apikey-generate .form-input {
-  flex: 1;
-  max-width: 280px;
-}
-
-.apikey-reveal {
-  background: var(--bg-primary);
-  border: 1px solid var(--accent-gold-dim);
-  border-radius: var(--radius-sm);
-  padding: 0.75rem 1rem;
-  margin-bottom: 0.75rem;
-}
-
-.apikey-reveal-warning {
-  font-size: 0.8rem;
-  color: var(--accent-gold);
-  margin-bottom: 0.5rem;
-  font-weight: 500;
-}
-
-.apikey-reveal-box {
-  display: flex;
-  gap: 0.5rem;
-  align-items: center;
-}
-
-.apikey-reveal-value {
-  flex: 1;
-  font-size: 0.78rem;
-  background: var(--bg-card);
-  padding: 0.4rem 0.6rem;
-  border-radius: var(--radius-sm);
-  word-break: break-all;
-  user-select: all;
-}
-
-.apikey-list {
-  margin-top: 1rem;
-  display: flex;
-  flex-direction: column;
-  gap: 0.5rem;
-}
-
-.apikey-item {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 0.6rem 0;
-  border-bottom: 1px solid var(--border-subtle);
-  gap: 0.75rem;
-}
-
-.apikey-item:last-child {
-  border-bottom: none;
-}
-
-.apikey-item.revoked {
-  opacity: 0.5;
-}
-
-.apikey-item-info {
-  display: flex;
-  flex-direction: column;
-  gap: 0.1rem;
-  min-width: 0;
-}
-
-.apikey-item-name {
-  font-size: 0.9rem;
-  font-weight: 500;
-}
-
-.apikey-item-meta {
-  font-size: 0.75rem;
-  color: var(--text-muted);
-}
-
-.revoked-badge {
-  font-size: 0.7rem;
-  padding: 0.15rem 0.5rem;
-  background: var(--bg-primary);
-  border-radius: var(--radius-full);
-  color: var(--text-muted);
 }
 
 .btn-danger {
@@ -1256,116 +491,5 @@ onMounted(() => {
     font-size: 0.78rem;
     padding: 0.5rem 0.6rem;
   }
-}
-
-.modal-overlay {
-  position: fixed;
-  inset: 0;
-  background: rgba(0, 0, 0, 0.7);
-  display: flex;
-  align-items: flex-start;
-  justify-content: center;
-  padding: 15vh 1rem;
-  z-index: 1000;
-}
-
-.modal-content {
-  width: 100%;
-}
-
-.modal-header {
-  padding: 1rem 1.25rem;
-  border-bottom: 1px solid var(--border-subtle);
-}
-
-.modal-body {
-  padding: 1.25rem;
-}
-
-/* Tag Manager */
-.tag-create-form {
-  display: flex;
-  flex-wrap: wrap;
-  gap: 0.5rem;
-  align-items: center;
-  margin: 1rem 0;
-}
-
-.tag-create-form .form-input {
-  flex: 1;
-  min-width: 150px;
-}
-
-.tag-color-picker {
-  display: flex;
-  gap: 0.3rem;
-  align-items: center;
-}
-
-.color-swatch {
-  width: 22px;
-  height: 22px;
-  border-radius: 50%;
-  border: 2px solid transparent;
-  cursor: pointer;
-  padding: 0;
-}
-
-.color-swatch.active {
-  border-color: var(--text-primary);
-  box-shadow: 0 0 0 2px var(--bg-card);
-}
-
-.color-swatch.sm {
-  width: 18px;
-  height: 18px;
-}
-
-.tag-error {
-  color: #ef4444;
-  font-size: 0.85rem;
-  margin-top: 0.25rem;
-}
-
-.tag-list {
-  margin-top: 1rem;
-  display: flex;
-  flex-direction: column;
-  gap: 0.5rem;
-}
-
-.tag-list-item {
-  display: flex;
-  align-items: center;
-  gap: 0.5rem;
-  padding: 0.5rem;
-  border: 1px solid var(--border-subtle);
-  border-radius: var(--radius-sm);
-  flex-wrap: wrap;
-}
-
-.tag-preview {
-  font-size: 0.8rem;
-  padding: 0.2rem 0.6rem;
-  border-radius: 9999px;
-  border: 1px solid;
-  flex-shrink: 0;
-}
-
-.tag-edit-input {
-  flex: 1;
-  min-width: 120px;
-}
-
-.tag-actions {
-  margin-left: auto;
-  display: flex;
-  gap: 0.25rem;
-}
-
-.empty-tags {
-  color: var(--text-secondary);
-  font-size: 0.85rem;
-  margin-top: 1rem;
 }
 </style>
